fix: better seo
